+++ src/sphere.rs
@@ -8,12 +8,12 @@ use std::rc::Rc;
 
 pub struct Sphere {
     pub center: Vec3,
-    pub radius: f32,
+    pub radius: f64,
     pub material: Rc<dyn Material>,
 }
 
 impl Sphere {
-    pub fn new(center: Vec3, radius: f32, material: Rc<dyn Material>) -> Sphere {
+    pub fn new(center: Vec3, radius: f64, material: Rc<dyn Material>) -> Sphere {
         Sphere {
             center,
             radius,
@@ -23,7 +23,7 @@ impl Sphere {
 }
 
 impl Hittable for Sphere {
-    fn hit(&self, ray: &Ray, t_min: f32, t_max: f32) -> Option<HitRecord> {
+    fn hit(&self, ray: &Ray, t_min: f64, t_max: f64) -> Option<HitRecord> {
         let oc = ray.origin - self.center;
         let a = ray.direction.squared_length();
         let half_b = oc.dot(&ray.direction);
@@ -43,7 +43,7 @@ impl Hittable for Sphere {
         return None;
     }
 
-    fn bounding_box(&self, t0: f32, t1: f32) -> Option<AABB> {
+    fn bounding_box(&self, t0: f64, t1: f64) -> Option<AABB> {
         let vec_radius = Vec3::new(self.radius, self.radius, self.radius);
         Some(AABB::new(
             self.center - vec_radius,
@@ -55,14 +55,14 @@ impl Hittable for Sphere {
 pub struct MovingSphere {
     pub center0: Vec3,
     pub center1: Vec3,
-    pub time0: f32,
-    pub time1: f32,
-    pub radius: f32,
+    pub time0: f64,
+    pub time1: f64,
+    pub radius: f64,
     pub material: Rc<dyn Material>,
 }
 
 impl MovingSphere {
-    pub fn new(c0: Vec3, c1: Vec3, t0: f32, t1: f32, r: f32, m: Rc<dyn Material>) -> MovingSphere {
+    pub fn new(c0: Vec3, c1: Vec3, t0: f64, t1: f64, r: f64, m: Rc<dyn Material>) -> MovingSphere {
         MovingSphere {
             center0: c0,
             center1: c1,
@@ -73,14 +73,14 @@ impl MovingSphere {
         }
     }
 
-    pub fn center(&self, time: f32) -> Vec3 {
+    pub fn center(&self, time: f64) -> Vec3 {
         self.center0
             + ((time - self.time0) / (self.time1 - self.time0)) * (self.center1 - self.center0)
     }
 }
 
 impl Hittable for MovingSphere {
-    fn hit(&self, ray: &Ray, t_min: f32, t_max: f32) -> Option<HitRecord> {
+    fn hit(&self, ray: &Ray, t_min: f64, t_max: f64) -> Option<HitRecord> {
         let oc = ray.origin - self.center(ray.time);
         let a = ray.direction.squared_length();
         let half_b = oc.dot(&ray.direction);
@@ -100,7 +100,7 @@ impl Hittable for MovingSphere {
         None
     }
 
-    fn bounding_box(&self, t0: f32, t1: f32) -> Option<AABB> {
+    fn bounding_box(&self, t0: f64, t1: f64) -> Option<AABB> {
         let vec_radius = Vec3::new(self.radius, self.radius, self.radius);
         let box0 = AABB::new(
             self.center(self.time0) - vec_radius,
